# BookManagementSystem/swagger.json
{
  "openapi": "3.1.0",
  "info": {
      "title": "BookManagement",
      "description": "API documentation for Book Management system",
      "version": "1.0.0",
      "contact": {}
  },
  "servers": [
      {
          "url": "http://localhost:3000",
          "description" : "Local Server"
      },
      {
        "url": "http://localhost:4000",
        "description" : "Production Server"
      }
  ],
  "paths": {
...
      "/api/category/getAllCategories": {
          "get": {
              "tags": [
                  "Category"
              ],
              "summary": "GetCategory",
              "description": "GetCategory",
              "operationId": "getcategory",
              "parameters": [
                  {
                      "name": "page",
                      "in": "query",
                      "schema": {
                          "type": "string",
                          "example": "2"
                      }
                  },
                  {
                      "name": "limit",
                      "in": "query",
                      "schema": {
                          "type": "string",
                          "example": "1"
                      }
                  }
              ],
              "requestBody": {
                  "content": {
                      "application/json": {
                          "examples": {
                              "GetCategory": {
                                  "value": ""
                              }
                          }
                      }
                  }
              },
              "responses": {
                  "200": {
                      "description": ""
                  }
              }
          }
      },
      "/api/category/getCategoryById": {
          "get": {
              "tags": [
                  "Category"
              ],
              "summary": "GetCategoryById",
              "description": "GetCategoryById",
              "operationId": "getcategorybyid",
              "requestBody": {
                  "content": {
                      "application/json": {
                          "schema": {
                              "type": "object",
                              "properties": {
                                  "id": {
                                      "type": "string",
                                      "example": "6638bd8f222f1a413897795fd"
                                  }
                              }
                          },
                          "examples": {
                              "GetCategoryById": {
                                  "value": {
                                      "id": "6638bd8f222f1a413897795fd"
                                  }
                              }
                          }
                      }
                  }
              },
              "responses": {
                  "200": {
                      "description": ""
                  }
              }
          }
      },
      "/api/category/searchCategories": {
          "get": {
              "tags": [
                  "Category"
              ],
              "summary": "SearchCategory",
              "description": "SearchCategory",
              "operationId": "searchcategory",
              "requestBody": {
                  "content": {
                      "application/json": {
                          "schema": {
                              "type": "object",
                              "properties": {
                                  "search": {
                                      "type": "string",
                                      "example": "Fiction"
                                  }
                              }
                          },
                          "examples": {
                              "SearchCategory": {
                                  "value": {
                                      "search": "Fiction"
                                  }
                              }
                          }
                      }
                  }
              },
              "responses": {
                  "200": {
                      "description": ""
                  }
              }
          }
      },
...
      "/api/books": {
          "get": {
              "tags": [
                  "Books"
              ],
              "summary": "GetBooks",
              "description": "GetBooks",
              "operationId": "getbooks",
              "parameters": [
                  {
                      "name": "page",
                      "in": "query",
                      "schema": {
                          "type": "string",
                          "example": "2"
                      }
                  },
                  {
                      "name": "limit",
                      "in": "query",
                      "schema": {
                          "type": "string",
                          "example": "1"
                      }
                  }
              ],
              "responses": {
                  "200": {
                      "description": ""
                  }
              }
          }
      },
...
      "/api/books/search": {
          "get": {
              "tags": [
                  "Books"
              ],
              "summary": "SearchBooks",
              "description": "SearchBooks",
              "operationId": "searchbooks",
              "parameters": [
                  {
                      "name": "q",
                      "in": "query",
                      "schema": {
                          "type": "string",
                          "example": "Mocking"
                      }
                  }
              ],
              "responses": {
                  "200": {
                      "description": ""
                  }
              }
          }
      },
      "/api/books/filter": {
          "get": {
              "tags": [
                  "Books"
              ],
              "summary": "FilterBooks",
              "description": "\nThis endpoint makes an HTTP GET request to filter books. The request does not include any specific parameters in the URL or request body. \n\n### Response\nThe response returned with a status code of 200 and a content type of application/json. The response body is an empty array [].\n\n#### JSON Schema\n```json\n{\n  \"type\": \"array\",\n  \"items\": {}\n}\n```\n",
              "operationId": "filterbooks",
              "parameters": [
                  {
                      "name": "query",
                      "in": "query",
                      "schema": {
                          "type": "string",
                          "example": "Philosophy"
                      }
                  },
                  {
                      "name": "minPrice",
                      "in": "query",
                      "schema": {
                          "type": "string",
                          "example": "300"
                      }
                  },
                  {
                      "name": "maxPrice",
                      "in": "query",
                      "schema": {
                          "type": "string",
                          "example": "600"
                      }
                  }
              ],
              "responses": {
                  "200": {
                      "description": ""
                  }
              },
              "security": [
                  {
                      "bearerAuth": []
                  }
              ]
          }
      },
      "/api/books/": {
          "post": {
              "tags": [
                  "Books"
              ],
              "summary": "AddBooks",
              "description": "AddBooks",
              "operationId": "addbooks",
              "requestBody": {
                  "content": {
                      "application/json": {
                          "schema": {
                              "type": "object",
                              "properties": {
                                  "ISBN": {
                                      "type": "string",
                                      "example": "9780143130727"
                                  },
                                  "author": {
                                      "type": "string",
                                      "example": "6639cc2b4fd35166e8a9df06"
                                  },
                                  "category": {
                                      "type": "string",
                                      "example": "6639d462f2d0d5ae0b6e3b36"
                                  },
                                  "description": {
                                      "type": "string",
                                      "example": "THE JAPANESE SECRET TO A LONG AND HAPPY LIFE"
                                  },
                                  "price": {
                                      "type": "number",
                                      "example": 399
                                  },
                                  "publishedYear": {
                                      "type": "number",
                                      "example": 1998
                                  },
                                  "title": {
                                      "type": "string",
                                      "example": "Ikigai"
                                  }
                              }
                          },
                          "examples": {
                              "AddBooks": {
                                  "value": {
                                      "ISBN": "9780143130727",
                                      "author": "6639cc2b4fd35166e8a9df06",
                                      "category": "6639d462f2d0d5ae0b6e3b36",
                                      "description": "THE JAPANESE SECRET TO A LONG AND HAPPY LIFE",
                                      "price": 399,
                                      "publishedYear": 1998,
                                      "title": "Ikigai"
                                  }
                              }
                          }
                      }
                  }
              },
              "responses": {
                  "200": {
                      "description": ""
                  }
              }
          }
      },
...
  },
  "components": {
      "securitySchemes": {
          "bearerAuth": {
              "type": "http",
              "scheme": "bearer"
          }
        },
...
  },
...
}